Update photonKernelVal
Merged 'dev' into 'photonKernelVal' with an additional update for photon dose calculation. The Update takes care of the extended kernel version (see photonPencilBeamKernelCalc).
Furthermore, implementation of the collimator rotation (for import process and recalculation) which takes account for possible collimator rotation of imported plans. Be aware, Collimator rotation is not yet implemented in GUI.